refactor(migrate): rename targets => destinations
